Add plate cutouts to the User.1 layer.
The intention is that this layer can be plotted as a Gerber and renamed
to be the EdgeCuts. That way, the plate cutouts _always_ match the
switch locations on the PCB.

Cutout dimensions are heavily inspired by (read: stolen from) the
awesome: http://builder.swillkb.com/

Switch cutouts are square for maximum stability. Stabilizers are
combined cherry/costar. Since this keyboard originally shipped with
costar stabilizers, it's nice to keep compatibility.

TODO: Add some extra margin around the edges of the plate so there
aren't any thin lines at the edges.
Double check the stab cutout dimensions with the Cherry MX datasheet.
Consider adding the original mounting hole under the F-row.